# Performance for EndpointMetadataCollection