--- a/test/Microsoft.AspNetCore.Mvc.Routing.Abstractions.Tests/EndpointMetadataCollectionTests.cs
+++ b/test/Microsoft.AspNetCore.Mvc.Routing.Abstractions.Tests/EndpointMetadataCollectionTests.cs
@@ -44,5 +44,98 @@ namespace Microsoft.AspNetCore.Routing
                 value => Assert.Equal(2, value),
                 value => Assert.Equal(3, value));
         }
+
+        [Fact]
+        public void GetMetadata_Match_ReturnsLastMatchingEntry()
+        {
+            // Arrange
+            var items = new object[]
+            {
+                new Metadata1(),
+                new Metadata2(),
+                new Metadata3(),
+            };
+
+            var metadata = new EndpointMetadataCollection(items);
+
+            // Act
+            var result = metadata.GetMetadata<IMetadata5>();
+
+            // Assert
+            Assert.Same(items[1], result);
+        }
+
+        [Fact]
+        public void GetMetadata_NoMatch_ReturnsNull()
+        {
+            // Arrange
+            var items = new object[]
+            {
+                new Metadata3(),
+                new Metadata3(),
+                new Metadata3(),
+            };
+
+            var metadata = new EndpointMetadataCollection(items);
+
+            // Act
+            var result = metadata.GetMetadata<IMetadata5>();
+
+            // Assert
+            Assert.Null(result);
+        }
+
+        [Fact]
+        public void GetOrderedMetadata_Match_ReturnsItemsInAscendingOrder()
+        {
+            // Arrange
+            var items = new object[]
+            {
+                new Metadata1(),
+                new Metadata2(),
+                new Metadata3(),
+            };
+
+            var metadata = new EndpointMetadataCollection(items);
+
+            // Act
+            var result = metadata.GetOrderedMetadata<IMetadata5>();
+
+            // Assert
+            Assert.Collection(
+                result,
+                i => Assert.Same(items[0], i),
+                i => Assert.Same(items[1], i));
+        }
+
+        [Fact]
+        public void GetOrderedMetadata_NoMatch_ReturnsEmpty()
+        {
+            // Arrange
+            var items = new object[]
+            {
+                new Metadata3(),
+                new Metadata3(),
+                new Metadata3(),
+            };
+
+            var metadata = new EndpointMetadataCollection(items);
+
+            // Act
+            var result = metadata.GetOrderedMetadata<IMetadata5>();
+
+            // Assert
+            Assert.Empty(result);
+        }
+
+        private interface IMetadata1 { }
+        private interface IMetadata2 { }
+        private interface IMetadata3 { }
+        private interface IMetadata4 { }
+        private interface IMetadata5 { }
+        private class Metadata1 : IMetadata1, IMetadata4, IMetadata5 { }
+        private class Metadata2 : IMetadata2, IMetadata5 { }
+        private class Metadata3 : IMetadata3 { }
+
     }
 }